refactor(product) #1: 불필요한 파일 삭제 및 enum 파일 이동
